fix: extract hand types/DB ops to lib to avoid cross-'use server' imports
Turbopack forbids importing between 'use server' files. Moved HAND_SIZE,
hand types, readHandDb(), and addBarToHandForPlayer() to src/lib/hand-service.ts
(no 'use server'). capture-bar and daily-charge now import from the lib;
hand.ts stays as the auth-gated server-action API wrapping the same logic.
